+++ modules/database/variables.tf
@@ -30,6 +30,7 @@ variable "availability_type" {
   default = "REGIONAL"
 }
 
+
 variable "maintenance_window_day" {
   description = "The day of week (1-7) for the master instance maintenance."
   type        = number
@@ -71,4 +72,10 @@ variable "force_ssl" {
   description = "Enforce SSL through the usage of the Cloud SQL Proxy (cloudsql://) in the DB connection string"
   type        = bool
   default     = false
+}
+
+variable "crypto_key" {
+  type        = string
+  default     = null
+  description = "Key used to encrypt and decrypt database."
 }