+++ modules/kms/variables.tf
@@ -8,3 +8,15 @@ variable "deletion_protection" {
   type        = bool
   default     = true
 }
+
+variable "key_location" {
+  type        = string
+  description = "Location where the KMS key will be created."
+  default = "global"
+}
+
+variable "bind_pubsub_service_to_kms_key" {
+  type        = bool
+  description = "Whether to bind the Pub/Sub service account to the KMS key for encrypter/decrypter access."
+  default = true
+}